+++ lib/ext/Assistant.dart
@@ -17,11 +17,13 @@
 import 'package:illinois/model/Assistant.dart';
 import 'package:illinois/model/Dining.dart';
 import 'package:rokwire_plugin/model/event2.dart';
+import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
 ///////////////////////////////
 /// Message
-extension MessageExt on Message {
+
+extension MessageUI on Message {
 
   List<dynamic>? get structElements {
     List<dynamic>? elements;
@@ -44,4 +46,56 @@ extension MessageExt on Message {
     }
     return elements;
   }
+
+  String get providerDisplayString =>
+    provider?.displayString ?? Localization().getStringEx('model.assistant.provider.unknown.label', 'Unknown');
+
+}
+
+extension AssistantProviderUI on AssistantProvider {
+
+  String get displayString {
+    switch (this) {
+      case AssistantProvider.google: return Localization().getStringEx('model.assistant.provider.google.label', 'Google');
+      case AssistantProvider.grok: return Localization().getStringEx('model.assistant.provider.grok.label', 'Grok');
+      case AssistantProvider.perplexity: return Localization().getStringEx('model.assistant.provider.perplexity.label', 'Perplexity');
+      case AssistantProvider.openai: return Localization().getStringEx('model.assistant.provider.openai.label', 'Illinois');
+    }
+  }
+
+  static AssistantProvider? fromCode(String? code) {
+    switch (code) {
+      case 'google_assistant': return AssistantProvider.google;
+      case 'grok_assistant': return AssistantProvider.grok;
+      case 'perplexity_assistant': return AssistantProvider.perplexity;
+      case 'openai_assistant': return AssistantProvider.openai;
+      default: return null;
+    }
+  }
+
+  static List<AssistantProvider>? listFromCodes(Iterable<String>? codes) {
+    if (codes != null) {
+      List<AssistantProvider> providers = <AssistantProvider>[];
+      for (String code in codes) {
+        AssistantProvider? provider = fromCode(code);
+        if (provider != null) {
+          providers.add(provider);
+        }
+      }
+      return providers;
+    }
+    else {
+      return null;
+    }
+  }
+}
+
+extension AssistantSettingsUI on AssistantSettings {
+
+  String? get localizedTermsText => _getTermsText(locale: _localeCode);
+  String? get localizedUnavailableText => _getUnavailableText(locale: _localeCode);
+
+  String? _getTermsText({required String locale}) => JsonUtils.stringValue(termsTextJson?[locale]) ;
+  String? _getUnavailableText({required String locale}) => JsonUtils.stringValue(unavailableTextJson?[locale]);
+  String get _localeCode => Localization().currentLocale?.languageCode ?? Localization().defaultLocale?.languageCode ?? 'en';
 }